+++ src/day2.h
@@ -6,7 +6,7 @@
 
 namespace day2
 {
-	std::pair<int, int> solve(const std::vector<std::string>& input);
+	std::pair<int64_t, int64_t> solve(const std::vector<std::string>& input);
