+++ src/day7.cpp
@@ -69,7 +69,7 @@ namespace day7
 		}
 	}
 
-	pair<int, int> solve(const vector<string>& input)
+	pair<int64_t, int64_t> solve(const vector<string>& input)
 	{
 		auto program = util::mapVector<string, int64_t>(util::split(input[0], ','));
