+++ src/day8.h
@@ -5,5 +5,5 @@
 
 namespace day8
 {
-	std::pair<int, int> solve(const std::vector<std::string>& input);
+	std::pair<int64_t, int64_t> solve(const std::vector<std::string>& input);
 }
